🔖 0.15.1 (#62)
* 🎨 [tcr.CloneResidency] Add indicators during running

* 🐛 [misc.File2Proc] Check if input file exists

* 🎨 [tcr.Immunarch] Allow cases for trackings and add mutaters for metadata

* 🎨 [tcr.CloneResidency] Allow multiple cases add mutaters for metadata

* ⬆️ Bump pipen-board to 0.8.0

* ⬆️ Bump pipen-report to 0.12.5 (to fix the pydantic error)

* 🔖 0.15.1

+++ biopipen/ns/misc.py
@@ -18,6 +18,10 @@ class File2Proc(Proc):
     script = """
         # in case of deadlink
         rm -f {{out.outfile | quote}}
+        if [[ ! -e {{in.infile | quote}} ]]; then
+            echo "File {{in.infile | quote}} does not exist." 1>&2
+            exit 1
+        fi
         ln -s {{in.infile | quote}} {{out.outfile | quote}}
     """
